Pump functions to new lib
