+++ src/main.c
@@ -13,11 +13,12 @@ TODO:
 #include <string.h>
 
 #include "hydro_pinout.h"
+#include "hydro_pump.h"
 #include "hydro_sensors.h"
 
 // drivers
+#include "driver/gpio.h"
 #include "driver/ledc.h"
-#include "driver\gpio.h"
 
 hydro_sensor_t sensors[] = {
     {
@@ -90,227 +91,6 @@ hydro_sensor_t sensors[] = {
     //         },
     // },
 };
-
-// void bme280_task(void *pvParameters)
-// {
-//     static const char *TAG = "BME280";
-
-//     bmp280_params_t params;
-//     bmp280_init_default_params(&params);
-//     bmp280_t dev;
-//     memset(&dev, 0, sizeof(bmp280_t));
-
-//     ESP_ERROR_CHECK(bmp280_init_desc(&dev, BMP280_I2C_ADDRESS_1, 0, HYDRO_PINOUT_I2C0_SDA, HYDRO_PINOUT_I2C0_SCL));
-
-//     ESP_ERROR_CHECK(bmp280_init(&dev, &params));
-
-//     bool bme280p = dev.id == BME280_CHIP_ID;
-//     ESP_LOGI(TAG, "BMP280: found %s", bme280p ? "BME280" : "BMP280");
-
-//     float pressure, temperature, humidity;
-
-//     while (1)
-//     {
-//         vTaskDelay(pdMS_TO_TICKS(5000));
-//         if (bmp280_read_float(&dev, &temperature, &pressure, &humidity) != ESP_OK)
-//         {
-//             ESP_LOGE(TAG, "Temperature/pressure reading failed");
-//             continue;
-//         }
-
-//         ESP_LOGI(TAG, "Pressure: %.2f Pa, Temperature: %.2f C, Humidity: %.2f%%", pressure, temperature, humidity);
-//     }
-// }
-
-// void bme680_task(void *pvParameters)
-// {
-//     static const char *TAG = "BME680";
-
-//     // vTaskDelay(pdMS_TO_TICKS(2000));
-
-//     // bme680_t sensor = sensors[1].sensor_obj.bme680;
-//     // memset(&sensor, 0, sizeof(bme680_t));
-
-//     // ESP_ERROR_CHECK(bme680_init_desc(&sensor, BME680_I2C_ADDR_0, 0, HYDRO_PINOUT_I2C0_SDA,
-//     HYDRO_PINOUT_I2C0_SCL));
-
-//     // // init the sensor
-//     // ESP_ERROR_CHECK(bme680_init_sensor(&sensor));
-
-//     // // Changes the oversampling rates to 4x oversampling for temperature
-//     // // and 2x oversampling for humidity. Pressure measurement is skipped.
-//     // bme680_set_oversampling_rates(&sensor, BME680_OSR_4X, BME680_OSR_NONE, BME680_OSR_2X);
-
-//     // // Change the IIR filter size for temperature and pressure to 7.
-//     // bme680_set_filter_size(&sensor, BME680_IIR_SIZE_7);
-
-//     // // Change the heater profile 0 to 200 degree Celsius for 100 ms.
-//     // bme680_set_heater_profile(&sensor, 0, 200, 100);
-//     // bme680_use_heater_profile(&sensor, 0);
-
-//     // // Set ambient temperature to 10 degree Celsius
-//     // bme680_set_ambient_temperature(&sensor, 10);
-
-//     // as long as sensor configuration isn't changed, duration is constant
-//     // uint32_t duration;
-//     // bme680_get_measurement_duration(&sensor, &duration);
-
-//     TickType_t last_wakeup = xTaskGetTickCount();
-
-//     // bme680_values_float_t values;
-//     hydro_data_t values;
-//     esp_err_t err;
-//     while (1)
-//     {
-//         err = read_sensor(TAG, &sensors[1], &values);
-//         if (err != ESP_OK)
-//             ESP_LOGE(TAG, "Error reading sensor: %d", err);
-//         else
-//         {
-//             ESP_LOGI(TAG, "Sensor: %.2f °C, %.2f %%, %.2f hPa, %.2f Ohm",
-//             values.data.temp_hum_press_gas.temperature_c,
-//                      values.data.temp_hum_press_gas.humidity, values.data.temp_hum_press_gas.pressure_pa,
-//                      values.data.temp_hum_press_gas.gas_resistance_ohm);
-//         }
-//         // // trigger the sensor to start one TPHG measurement cycle
-//         // if (bme680_force_measurement(&sensor) == ESP_OK)
-//         // {
-//         //     // passive waiting until measurement results are available
-//         //     vTaskDelay(duration);
-
-//         //     // get the results and do something with them
-//         //     if (bme680_get_results_float(&sensor, &values) == ESP_OK)
-//         //         ESP_LOGI(TAG, "Sensor: %.2f °C, %.2f %%, %.2f hPa, %.2f Ohm", values.temperature, values.humidity,
-//         //                  values.pressure, values.gas_resistance);
-//         // }
-//         // passive waiting until 5 seconds is over
-//         vTaskDelayUntil(&last_wakeup, pdMS_TO_TICKS(5000));
-//     }
-// }
-
-// void gpio_task(void *pvParameters)
-// {
-
-//     static const char *TAG = "GPIO";
-//     gpio_set_direction(HYDRO_PINOUT_PUMP, GPIO_MODE_OUTPUT);
-
-//     ESP_LOGI(TAG, "Pump is running");
-
-//     while (1)
-//     {
-//         gpio_set_level(HYDRO_PINOUT_PUMP, 1);
-//         vTaskDelay(pdMS_TO_TICKS(2000));
-//         gpio_set_level(HYDRO_PINOUT_PUMP, 0);
-//         vTaskDelay(pdMS_TO_TICKS(2000));
-//     }
-// }
-
-// void pwm_task(void *pvParameters)
-// {
-//     static const char *TAG = "PWM";
-
-//     // Configure the PWM timer
-//     ledc_timer_config_t ledc_timer = {.speed_mode = LEDC_LOW_SPEED_MODE,
-//                                       .timer_num = LEDC_TIMER_0,
-//                                       .duty_resolution = LEDC_TIMER_13_BIT,
-//                                       .freq_hz = 5000,
-//                                       .clk_cfg = LEDC_AUTO_CLK};
-//     ESP_ERROR_CHECK(ledc_timer_config(&ledc_timer));
-
-//     // Configure the PWM channel
-//     ledc_channel_config_t ledc_channel = {.speed_mode = LEDC_LOW_SPEED_MODE,
-//                                           .channel = LEDC_CHANNEL_0,
-//                                           .timer_sel = LEDC_TIMER_0,
-//                                           .intr_type = LEDC_INTR_DISABLE,
-//                                           .gpio_num = HYDRO_PINOUT_PWM,
-//                                           .duty = 0,
-//                                           .hpoint = 0};
-//     ESP_ERROR_CHECK(ledc_channel_config(&ledc_channel));
-
-//     int duty = 0;
-//     int direction = 1;
-
-//     while (1)
-//     {
-//         ESP_LOGI(TAG, "Setting PWM duty to %d", duty);
-//         ESP_ERROR_CHECK(ledc_set_duty(ledc_channel.speed_mode, ledc_channel.channel, duty));
-//         ESP_ERROR_CHECK(ledc_update_duty(ledc_channel.speed_mode, ledc_channel.channel));
-
-//         duty += direction * 2000;
-//         if (duty >= 8191)
-//         {
-//             direction = -direction;
-//             duty = 8191;
-//         }
-//         else if (duty <= 0)
-//         {
-//             direction = -direction;
-//             duty = 0;
-//         }
-
-//         vTaskDelay(pdMS_TO_TICKS(1000));
-//     }
-// }
-
-bool pump_slow_start(ledc_channel_config_t *ledc_channel)
-{
-    bool noerror = true;
-    esp_err_t err = ESP_OK;
-
-    for (int i = 0; i < 8191; i += 1000)
-    {
-        if (i > 8191) i = 8191;
-
-        err = ledc_set_duty(ledc_channel->speed_mode, ledc_channel->channel, i);
-        if (err != ESP_OK)
-        {
-            ESP_LOGE("PUMP SLOW START", "Error setting duty: %d", err);
-            noerror = false;
-            break;
-        }
-
-        err = ledc_update_duty(ledc_channel->speed_mode, ledc_channel->channel);
-        if (err != ESP_OK)
-        {
-            ESP_LOGE("PUMP SLOW START", "Error updating duty: %d", err);
-            noerror = false;
-            break;
-        }
-        vTaskDelay(pdMS_TO_TICKS(100));
-    }
-
-    return noerror;
-}
-
-bool pump_slow_stop(ledc_channel_config_t *ledc_channel)
-{
-    bool noerror = true;
-    esp_err_t err = ESP_OK;
-
-    for (int i = 8191; i >= 0; i -= 1000)
-    {
-        if (i < 0) i = 0;
-
-        err = ledc_set_duty(ledc_channel->speed_mode, ledc_channel->channel, i);
-        if (err != ESP_OK)
-        {
-            ESP_LOGE("PUMP SLOW START", "Error setting duty: %d", err);
-            noerror = false;
-            break;
-        }
-
-        err = ledc_update_duty(ledc_channel->speed_mode, ledc_channel->channel);
-        if (err != ESP_OK)
-        {
-            ESP_LOGE("PUMP SLOW START", "Error updating duty: %d", err);
-            noerror = false;
-            break;
-        }
-        vTaskDelay(pdMS_TO_TICKS(100));
-    }
-
-    return noerror;
-}
 
 void pump_task(void *pvParameters)
 {
@@ -411,9 +191,6 @@ void app_main(void)
 
     printf("Hello world!\n");
 
-    // xTaskCreatePinnedToCore(gpio_task, "gpio-example", configMINIMAL_STACK_SIZE * 8, NULL, 5, NULL, APP_CPU_NUM);
-    // xTaskCreatePinnedToCore(pwm_task, "pwm-example", configMINIMAL_STACK_SIZE * 8, NULL, 5, NULL, APP_CPU_NUM);
-
     err = init_all_sensors(TAG, sensors, sizeof(sensors) / sizeof(hydro_sensor_t));
     if (err == ESP_ERR_NOT_FINISHED)
     {
@@ -428,6 +205,4 @@ void app_main(void)
     xTaskCreatePinnedToCore(sensors_task, "sensors-task", configMINIMAL_STACK_SIZE * 8, NULL, 5, NULL, APP_CPU_NUM);
 
     xTaskCreatePinnedToCore(pump_task, "pump-task", configMINIMAL_STACK_SIZE * 8, NULL, 5, NULL, APP_CPU_NUM);
-
-    // xTaskCreatePinnedToCore(bme680_task, "bme680-example", configMINIMAL_STACK_SIZE * 8, NULL, 5, NULL, APP_CPU_NUM);
 }
